Create QeApp class as the entry point of the app
